speech-assistant-openai-realtime-api-python/requirements.txt
fastapi 
uvicorn 
python-dotenv 
websockets 
twilio
pydub
openai-whisper